Update UriKind to Relative for birb command
The birb command has been spitting out that it is unavailable for
quite a while now. Doing a little digging, I found that the file
key provided by the API is merely a file, not a full URI. This
caused `UriKind.Absolute` to fail the check in `IsWellFormedString`
This update adjusts to relative so as to not change any logic
going on while allowing responses to be provided once again.