Finally upgrade dependencies + Fix style warnings
Floofbot has been running on a wildly old version
of Discord.Net for a good while now. Let's fix
that! I've also went through and updated what was
required in Discord.Addons.Interactive (including
the license field, which wasn't properly set),
solved the issues that caused in the primary
project, and while I was at it I updated styles
and provided suppressions for the present issues
to ensure code analysis doesn't give messages.

All of this because I intended to add a
command to pull guild avatars. Huh.